+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ Changelog for the **SODAR Core** Django app package. Loosely follows the
 Note that the issue IDs here refer to ones in the private CUBI GitLab.
 
 
+Unreleased
+==========
+
+Fixed
+-----
+
+- **Projectroles**
+    - Transferring project ownership to inherited owner not allowed (#534)
+
+
 v0.8.1 (2020-04-24)
 ===================